feat: make Template2 work with images from payload
